wokred in MySQLCategoriesDao class

+++ src/main/java/org/yearup/controllers/CategoriesController.java
@@ -4,6 +4,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 import org.yearup.data.CategoryDao;
 import org.yearup.data.ProductDao;
 import org.yearup.models.Category;
@@ -11,10 +12,9 @@ import org.yearup.models.Product;
 
 import java.util.List;
 
-// http://localhost:8080/categories
 
 @RestController
-@RequestMapping("categories")
+@RequestMapping("/categories")
 @CrossOrigin
 public class CategoriesController {
     private final CategoryDao categoryDao;
@@ -38,29 +38,65 @@ public class CategoriesController {
         return categoryDao.getById(idNumber);
     }
 
-    @GetMapping("{categoryId}/products")
+    @GetMapping("/products")
     public List<Product> getProductsById(@PathVariable int categoryId) {
-        return getProductsById(categoryId);
+        try
+        {
+            var cat = categoryDao.getById(categoryId);
+
+            if(cat == null)
+                throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+
+            return productDao.search(categoryId, null, null, null); //might be wrong, ask Ben
+        }
+        catch(Exception ex)
+        {
+            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Oops... our bad.");
+        }
     }
 
     @GetMapping("")
     @PreAuthorize("hasRole('ROLE_ADMIN')")
     @ResponseStatus(value = HttpStatus.CREATED)
     public Category addCategory(@RequestBody Category category) {
-        return categoryDao.create(category);
+        try{
+            return categoryDao.create(category);
+        }
+        catch (Exception e) {
+            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Oops... our bad.");
+        }
+
     }
 
 
     @GetMapping("{categoryId}")
     @PreAuthorize("hasRole('ROLE_ADMIN')")
     public void updateCategory(@PathVariable int categoryId, @RequestBody Category category) {
-        categoryDao.update(categoryId, category);
+        try{
+            categoryDao.update(categoryId, category);
+        }
+        catch(Exception e)
+        {
+            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Oops... our bad.");
+        }
+
     }
-    
+
     @GetMapping("{categoryId}")
     @PreAuthorize("hasRole('ROLE_ADMIN')")
     @ResponseStatus(value = HttpStatus.NO_CONTENT)
     public void deleteCategory(@PathVariable int id) {
-        categoryDao.delete(id);
+        try{
+            var cat = categoryDao.getById(id);
+
+            if (cat == null)
+                throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+
+            categoryDao.delete(id);
+        }
+        catch (Exception e) {
+            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Oops... our bad.");
+        }
+
     }
 }
